+++ .env.example
@@ -4,4 +4,3 @@ DB_HOST=
 DB_USER=
 DB_PASS=
 DB_DATABASE=
-SECRET=
